requirements.txt
python-dotenv~=1.1.0
discord.py~=2.5.2
requests~=2.32.3
Flask~=3.1.1
table2ascii~=1.1.3
google-api-python-client~=2.170.0
google-auth~=2.40.2
google-auth-oauthlib~=1.2.2
requests~=2.32.3
